moved ros logging stuff to smarts.ros.logging formatting
