+++ smarts/ros/src/smarts_ros/scripts/ros_driver.py
@@ -42,7 +42,7 @@ from smarts.core.utils.math import (
     yaw_from_quaternion,
 )
 from smarts.core.vehicle import VehicleState
-from smarts.ros import log_everything_to_ROS
+from smarts.ros.loggin import log_everything_to_ROS
 from smarts.sstudio.types import MapSpec
 from smarts.zoo import registry
